pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
    "pre-commit>=3.4.0",
    "pyupgrade>=3.10.1",
]

[project]
name = "subtitler"
version = "0.1.0"
description = ""
authors = [
    { name = "SEBROVATER", email = "sebrovskiy.k@gmail.com" },
]
dependencies = [
    "deepgram-sdk>=2.11.0",
    "pyaudio>=0.2.13",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
